feat: redo player w cleaned up animations
