+++ src/examples/ReactThreeJs/HexGrid/Models/Player/index.ts
@@ -1,0 +1,2 @@
+export * from "./Player";
+export * from "./PlayerScene";
