--- src/examples/ReactThreeJs/HexGrid/Models/PlayerArcher/index.ts
@@ -1,2 +1,0 @@
-export * from "./PlayerArcher";
-export * from "./PlayerArcherScene";
