+++ src/examples/ReactThreeJs/HexGrid/Models/index.ts
@@ -6,4 +6,4 @@ export * from "./OverlayHighlight";
 export * from "./OverlayHighlightOutline";
 export * from "./OverlaySphere";
 export * from "./OverlayText";
-export * from "./PlayerArcher";
+export * from "./Player";
